내 명함 정보 조회(멀티건) 카운트 단위 테스트

+++ src/main/java/egovframework/com/cop/ncm/service/impl/NcrdManageDAO.java
@@ -160,6 +160,6 @@ public class NcrdManageDAO extends EgovComAbstractDAO {
      * @throws Exception
      */
     public int selectMyNcrdItemListCnt(NameCardVO ncrdVO) throws Exception {
-	return (Integer)selectOne("NcrdManageDAO.selectMyNcrdItemListCnt", ncrdVO);
+        return selectOne("NcrdManageDAO.selectMyNcrdItemListCnt", ncrdVO);
     }
 }
